Smallest change that I could make that will *actually* allow replacing the model with another type. Failed approach 1: It's not possible to change the factory type because that's never exposed to classes derived from StoreBasketController, (and even if it was, the BasketControllerBase type is declared as using exactly <StoreBasketModel>). (As a side note, the factory actually has some business logic in it, which can't be ideal)
Failed approach 2: It's not possible to simply copy/paste StoreBasketModel to own class and inherit directly from BasketControllerBase, because we then lose functionality to items marked "internal" to merchello core that StoreBasketController uses.

Successful approach: Make StoreBasketController a generic accepting a TBasketModel.  This to some extent ruins the point of having a factory, which was to use IOC to allow callers to specify different ways of creating something matching the proper interface.  It looks like the design here was trying to allow this kind of factory substitution, but in practice it seemed to be too locked up to actually use.